Chore: browser folder components cleanup
