+++ src/pages/browser/detail/FolderDetail.jsx
@@ -1,51 +1,52 @@
 import { useSelector } from 'react-redux'
-import AttributeTable from '/src/containers/attributeTable'
-import { getTaskTypeIcon } from '/src/utils'
-import { TagsField } from '/src/containers/fieldFormat'
 import { Panel } from '@ynput/ayon-react-components'
+import Thumbnail from '/src/containers/thumbnail'
+import AttributeTable from '/src/containers/attributeTable'
+import { getFolderTypeIcon } from '/src/utils'
+import { TagsField } from '/src/containers/fieldFormat'
 import { useUpdateEntitiesDetailsMutation } from '/src/services/updateEntitiesDetails'
 import { useGetEntitiesDetailsQuery } from '/src/services/getEntitiesDetails'
-import StatusSelect from '../../components/status/statusSelect'
+import StatusSelect from '/src/components/status/statusSelect'
 import usePubSub from '/src/hooks/usePubSub'
 
-const TaskDetail = () => {
+const FolderDetail = () => {
   const context = useSelector((state) => ({ ...state.context }))
   const projectName = context.projectName
-  const focusedTasks = context.focused.tasks
-  const taskId = focusedTasks.length === 1 ? focusedTasks[0] : null
+  const focusedFolders = context.focused.folders
+  const folderId = focusedFolders.length === 1 ? focusedFolders[0] : null
 
   // GET RTK QUERY
   const {
-    data: tasksData,
+    data: foldersData = [],
     isError,
     isLoading,
     refetch,
   } = useGetEntitiesDetailsQuery(
     {
       projectName,
-      ids: [taskId],
-      type: 'task',
+      ids: [folderId],
+      type: 'folder',
     },
-    { skip: !taskId },
+    { skip: !folderId },
   )
 
   // PUBSUB HOOK
-  usePubSub('entity.task', refetch, focusedTasks)
+  usePubSub('entity.folder', refetch, focusedFolders)
 
   // PATCH FOLDERS DATA
-  const [updateTask] = useUpdateEntitiesDetailsMutation()
+  const [updateFolder] = useUpdateEntitiesDetailsMutation()
 
   if (isLoading) return 'loading..'
 
-  if (isError) return 'ERROR: Soemthing went wrong...'
+  if (isError) return 'ERROR: Something went wrong...'
 
   const handleStatusChange = async (value, entity) => {
     try {
       const patches = [{ ...entity, status: value }]
 
-      const payload = await updateTask({
+      const payload = await updateFolder({
         projectName,
-        type: 'task',
+        type: 'folder',
         data: { status: value },
         patches,
       }).unwrap()
@@ -56,49 +57,49 @@ const TaskDetail = () => {
     }
   }
 
-  if (focusedTasks.length > 1) {
+  const folder = foldersData[0]?.node
+
+  if (!folder) return null
+
+  if (focusedFolders.length > 1) {
     return (
       <Panel>
-        <span>{focusedTasks.length} tasks selected</span>
+        <span>{focusedFolders.length} focusedFolders selected</span>
       </Panel>
     )
   }
-
-  const task = tasksData[0].node
 
   return (
     <Panel>
       <h3>
         <span className="material-symbols-outlined" style={{ verticalAlign: 'bottom' }}>
-          {getTaskTypeIcon(task.taskType)}
+          {getFolderTypeIcon(folder.folderType)}
         </span>
-        <span style={{ marginLeft: 10 }}>{task.label || task.name}</span>
+        <span style={{ marginLeft: 10 }}>{folder.label || folder.name}</span>
       </h3>
+      <Thumbnail projectName={projectName} entityType="folder" entityId={folderId} />
       <AttributeTable
-        entityType="task"
-        data={task.attrib}
+        entityType="folder"
+        data={folder.attrib}
         additionalData={[
-          { title: 'Task Type', value: task.taskType },
+          { title: 'Folder type', value: folder.folderType },
           {
             title: 'Status',
             value: (
               <StatusSelect
-                value={task.status}
+                value={folder.status}
                 statuses={context.project.statuses}
                 align={'right'}
-                onChange={(v) => handleStatusChange(v, task)}
+                onChange={(v) => handleStatusChange(v, folder)}
               />
             ),
           },
-          { title: 'Tags', value: <TagsField value={task.tags} /> },
-          {
-            title: 'Assignees',
-            value: task.assignees?.length ? task.assignees.join(', ') : '-',
-          },
+          { title: 'Tags', value: <TagsField value={folder.tags} /> },
+          { title: 'Path', value: folder.path },
         ]}
       />
     </Panel>
   )
 }
 
-export default TaskDetail
+export default FolderDetail
